move helper functions to helper file for reuse
